WIP: Fixing the current placement system
